[MOD]-> int id TO Long id

+++ src/main/java/model/service/AirlineService.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 
 public interface AirlineService {
     AirlineDto save(AirlineDto airline);
-    Optional<AirlineDto> findById(int id);
-    Optional<AirlineDto> update(int id, AirlineDto airline);
+    Optional<AirlineDto> findById(Long id);
+    Optional<AirlineDto> update(Long id, AirlineDto airline);
     List<AirlineDto> findAll();
     List<AirlineDto> findByName(String name);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
